refactor: extract helpers for NostrFetcher to separate module
